FDB-storage-app/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
azure-common==1.1.28
azure-core==1.29.5
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-nspkg==3.0.2
azure-storage-blob==12.19.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
distro==1.8.0
fastapi==0.104.1
fastapi-health==0.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
isodate==0.6.1
numpy==1.26.2
openai==0.27.4
pandas==2.1.3
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.24.0.post1
